+++ src/repositories/ActiveMatchRepository/ActiveMatchRepository.ts
@@ -1,9 +1,8 @@
 import { ActiveMatch, Prisma, PrismaClient } from "@prisma/client";
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime";
 import { Team } from "../../types/common";
-import { splitArray } from "../../utils/splitArrays";
 import LeaderboardRepository from "../LeaderboardRepository";
-import { generateRandomId } from "../../utils";
+import { generateRandomId, splitArray } from "../../utils";
 import { ActiveMatchTeams, NewActiveMatchInput, PlayerInActiveMatch, UpdatePlayerInActiveMatchInput } from "./types";
 
 export class ActiveMatchRepository {
