+++ src/services/TeamAssignmentService.ts
@@ -44,7 +44,7 @@ export function createRandomTeams(ballchasers: ReadonlyArray<PlayerInQueue>): Ar
 }
 
 export async function setCaptains(ballChasers: ReadonlyArray<PlayerInQueue>): Promise<ReadonlyArray<PlayerInQueue>> {
-  // 🐧 Can we add a test that covers this, please?
+  // 🐧 Yes
   const sortedBallChaser = ballChasers.slice().sort((o, b) => b.mmr - o.mmr);
 
   //We are going to set the Blue Captain (firt person to choose) as the second highest MMR.
